Changed the file references from [uid].js to [userToken].js, and the logic in the API calls to utilize the user token which is passed via the API calls, instead of the user UID which was passed around.
